♻️ refactor: 검색 페이지 로직 분리
